Autodrive over bridges
Allows autodrive to find paths when driving up/down ramps and over
bridges.

This is done by updating `compute_obstacles` so that when we check for
whether a point is an obstacle or not, we also remember which points
were up or down ramps. Each such ramp point leads to a different zlevel.
We then iterate over those points and do a flood-fill starting at those
points on the other zlevel. Each point found by the flood-fill is then
re-evaluated for whether it is an obstacle also on the other zlevel.

During this process, we also remember which zlevel is the ground. For
example, when driving up a ramp, most points will be zlevel=0, but the
points on the ramp and bridge will have its ground at zlevel=1. The
existing pathfinder collission checks are updated to use this
groundlevel z when determining if we are about to hit an obstacle.